requirements.txt
asgiref==3.8.1
Django==5.0.4
django-appconf==1.0.6
django-classy-tags==4.1.0
django-imagekit==5.0.0
django-taggit==5.0.1
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.5.0
pilkit==3.0
pillow==10.3.0
psycopg2-binary==2.9.9
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
wheel==0.43.0
